+++ Frends.Smb.CreateDirectory/Frends.Smb.CreateDirectory/Definitions/AuthenticationMode.cs
@@ -1,0 +1,17 @@
+﻿namespace Frends.Smb.CreateDirectory.Definitions;
+
+/// <summary>
+/// Specifies authentication mode.
+/// </summary>
+public enum AuthenticationMode
+{
+    /// <summary>
+    /// NTLM authentication protocol.
+    /// </summary>
+    Ntlm,
+
+    /// <summary>
+    /// Kerberos authentication protocol.
+    /// </summary>
+    Kerberos,
+}